TT-5683 Publishing requires Paratext
- no Bible Id or publishing rows enabled without Paratext
- allowed editing publishing content as non-Paratext user
