+++ src/store/localization/exported-strings-name.json
@@ -1,1 +1,1 @@
-{ "stringsName": "strings5f82b6ff.json" }
+{ "stringsName": "strings5bb4c914.json" }
