+++ src/utils/index.ts
@@ -73,3 +73,4 @@ export * from './stringAvatar';
 export * from './ReactIsInDevelopmentMode';
 export * from './isUrl';
 export * from './useWaitForRemoteQueue';
+export * from './useHasParatext'
